requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contextlib2==21.6.0
cryptography==44.0.1
cssbeautifier==1.15.1
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-ninja==1.3.0
django-ninja-extra==0.22.3
django-ninja-jwt==5.3.5
django-unfold==0.49.1
djlint==1.36.4
EditorConfig==0.17.0
injector==0.22.0
jsbeautifier==1.15.1
json5==0.10.0
pathspec==0.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
